update .ipynb and .py, add data_loader.py
